test(pogod): root test registries at a bindable socket dir (mg-ef80)
Making an unbindable attach socket fatal to the spawn surfaced that 90 of
this package's 311 tests had never bound one. They build their Registry on
t.TempDir(), which on darwin returns a /var/folders/... path that exceeds
sun_path on its own — so the bind failed, the failure was only a log line,
and the tests passed while silently never exercising attach. The few tests
that did care already worked around it with a short /tmp dir.

Promote that workaround to the package-wide default: shortSocketDir moves
to testsupport_test.go and every test registry roots there. Production was
never affected — pogod takes its socket dir from config.AgentSocketDir,
which falls back to a short hashed path when POGO_HOME is too deep.

Also renames the agents in TestRestartOnCrashFlagDrivesBranching, which
derived them from its subtest names and so exceeded MaxAgentNameLen, and
gives cmd/pogod's stall-nudger tests the same helper.

Docs + CHANGELOG: MaxAgentNameLen is now a promise, not a ceiling.
