+++ internal/agent/coordinator_test.go
@@ -226,7 +226,7 @@ func TestStartCrewAgentCoordinatorRename(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	reg, err := NewRegistry(filepath.Join(tmpHome, "sockets"))
+	reg, err := NewRegistry(shortSocketDir(t))
 	if err != nil {
 		t.Fatalf("NewRegistry: %v", err)
 	}
